requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.10.3
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
beanie==1.26.0
pydantic-settings==2.5.2
passlib==1.7.4
email-validator==2.2.0
certifi==2024.8.30
aiosmtplib==4.0.2
black==25.9.0
requests==2.32.5
user-agents==2.2.0
google-genai==1.41.0
pinecone-client==6.0.0